split DataTable.jsx in multiple Components
