+++ src/utils/dataTable/TableCells.jsx
@@ -16,20 +16,21 @@ export default function getTableCellChild(type, editable, cellData, elementPrope
     switch (type) {
         case cellTypes.TEXT:
             if (editable) {
-                return <Input value={cellData}/>
+                return <Input value={cellData} style={elementProperties.style}/>
             } else {
-                return <Typography>{cellData}</Typography>
+                return <Typography style={elementProperties.style}>{cellData}</Typography>
             }
         case cellTypes.NUMBER:
             if (editable) {
-                return <Input value={cellData} type='number'/>
+                return <Input value={cellData} type='number' style={elementProperties.style}/>
             } else {
-                return <Typography>{cellData}</Typography>
+                return <Typography style={elementProperties.style}>{cellData}</Typography>
             }
         case cellTypes.BOOL:
             if (editable) {
                 return <Checkbox
                     checked={cellData}
+                    style={elementProperties.style}
                 />
             } else {
                 //TODO: improt FA-Icons
@@ -40,19 +41,20 @@ export default function getTableCellChild(type, editable, cellData, elementPrope
                 elementProperties.onMouseEnter(event, cellData)
             }} onMouseLeave={elementProperties.onMouseLeave}/>
         case cellTypes.CURRENCY:
-            return <Input value={cellData}
+            return <Input value={cellData} style={elementProperties.style}
                           endAdornment={<InputAdornment position="end">{elementProperties.currency}</InputAdornment>}/>
         case cellTypes.SELECTOR:
             return <TextField
                 value={cellData}
-                select>
+                select
+                style={elementProperties.style}>
                 {elementProperties.selectorOptions.map(option => {
                     return <MenuItem key={option.value} value={option.value}>
                         {option.label}
                     </MenuItem>
                 })}</TextField>
         default:
-            return <Typography>{cellData}</Typography>
+            return <Typography style={elementProperties.style}>{cellData}</Typography>
 
     }
